Stock

+++ src/modules/products/repositories/stock.repository.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { Model } from 'mongoose';
+import { Model, Types } from 'mongoose';
 import { InjectModel } from '@nestjs/mongoose';
 import { Stock } from '../entities/stock.entity';
 import { CreateStockDto } from '../dto/create-stock.dto';
@@ -15,8 +15,12 @@ export class StockRepository {
     return createdStock.save();
   }
 
-  findAllByProductId(productId: string): Promise<Stock[]> {
-    return this.stockModel
+  async findAllByProductId(productId: string): Promise<Stock[]> {
+    if (!Types.ObjectId.isValid(productId)) {
+      return [];
+    }
+
+    return await this.stockModel
       .find({
         productId: productId,
       })
@@ -24,11 +28,18 @@ export class StockRepository {
   }
 
   findAll(): Promise<Stock[]> {
-    return this.stockModel.find().exec();
+    return this.stockModel
+      .find()
+      .populate<{ product: Product }>('productId')
+      .exec();
   }
 
-  findOne(id: string): Promise<Stock | null> {
-    return this.stockModel
+  async findOne(id: string): Promise<Stock | null> {
+    if (!Types.ObjectId.isValid(id)) {
+      return null;
+    }
+
+    return await this.stockModel
       .findOne({
         _id: id,
       })
@@ -37,6 +48,10 @@ export class StockRepository {
   }
 
   async delete(id: string): Promise<void> {
+    if (!Types.ObjectId.isValid(id)) {
+      throw new Error('Invalid parameter');
+    }
+
     const result = await this.stockModel
       .deleteOne({
         _id: id,
@@ -46,5 +61,17 @@ export class StockRepository {
     if (result.deletedCount == 0) {
       throw new Error('Error on delete');
     }
+  }
+
+  async removeAllByProductId(productId: string): Promise<void> {
+    if (!Types.ObjectId.isValid(productId)) {
+      throw new Error('Invalid parameter');
+    }
+
+    await this.stockModel
+      .deleteMany({
+        productId: productId,
+      })
+      .exec();
   }
 }
